+++ src/nn.rs
@@ -17,7 +17,7 @@ impl VectorModel {
         let jj = self.in_vecs.len_of(Axis(1));
         for i in 0..ii {
             for j in 0..jj {
-                let mut v = self.in_vecs.slice_mut(s![i as usize, j, ..]);
+                let mut v = self.in_vecs.slice_mut(s![i, j, ..]);
                 let n = norm_l2(&v);
                 v.iter_mut().for_each(|e| *e /= n );
             }
@@ -30,11 +30,11 @@ pub fn sim(vm: &VectorModel,
             head_id1: usize, senseno1: usize,
             head_id2: usize, senseno2: usize) -> f32
 {
-    let qvec_r1 = vm.in_vecs.slice(s![head_id1 as usize, senseno1, ..]);
+    let qvec_r1 = vm.in_vecs.slice(s![head_id1, senseno1, ..]);
     let qnorm1 = norm_l2(&qvec_r1);
     let qvec1 = qvec_r1.mapv(|v| v / qnorm1);
 
-    let qvec_r2 = vm.in_vecs.slice(s![head_id2 as usize, senseno2, ..]);
+    let qvec_r2 = vm.in_vecs.slice(s![head_id2, senseno2, ..]);
     let qnorm2 = norm_l2(&qvec_r2);
     let qvec2 = qvec_r2.mapv(|v| v / qnorm2);
 
@@ -45,8 +45,8 @@ pub fn sim_normed(vm: &VectorModel,
             head_id1: usize, senseno1: usize,
             head_id2: usize, senseno2: usize) -> f32
 {
-    let qvec1 = vm.in_vecs.slice(s![head_id1 as usize, senseno1, ..]);
-    let qvec2 = vm.in_vecs.slice(s![head_id2 as usize, senseno2, ..]);
+    let qvec1 = vm.in_vecs.slice(s![head_id1, senseno1, ..]);
+    let qvec2 = vm.in_vecs.slice(s![head_id2, senseno2, ..]);
     qvec1.dot(&qvec2)
 }
 
@@ -55,7 +55,7 @@ pub fn nearest(vm: &VectorModel, head_id: usize, senseno: usize, top_k: usize, m
     let ii = vm.in_vecs.len_of(Axis(0));
     let jj = vm.in_vecs.len_of(Axis(1));
 
-    let qvec_r = vm.in_vecs.slice(s![head_id as usize, senseno, ..]);
+    let qvec_r = vm.in_vecs.slice(s![head_id, senseno, ..]);
     let qnorm = norm_l2(&qvec_r);
     let qvec = qvec_r.mapv(|v| v / qnorm);
